Adds webworker
