+++ web/.eslintrc.js
@@ -33,6 +33,9 @@ module.exports = {
     "semi": [
       "error",
       "always"
+    ],
+    "no-var": [
+      "error"
     ]
   }
 };
