+++ web/src/javascript/app.js
@@ -1,17 +1,23 @@
-import {draw as drawGame, fetchState} from "./game";
 import {drawLeaderboard} from "./gameLeaderboard";
-import {createGameCanvas} from "./drawMap";
+import {drawGame, createGameCanvas} from "./game";
 
 const gameCanvas = createGameCanvas();
 let currentState;
 
+(new Worker("javascript/gameWebWorker.bundle.js")).onmessage = message => {
+  currentState = message;
+};
+
 async function updateLoop() {
-  currentState = await fetchState(0);
-  drawGame(currentState, gameCanvas);
+  if(currentState) {
+    drawGame(currentState, gameCanvas);
+  }
 }
 
 function updateLeaderBoard() {
-  drawLeaderboard(currentState);
+  if(currentState) {
+    drawLeaderboard(currentState);
+  }
 }
 
 setInterval(updateLoop, 1000/24);
