+++ web/src/javascript/gameWebWorker.js
@@ -1,0 +1,7 @@
+import {fetchState} from "./network";
+
+async function updateLoop() {
+  postMessage(await fetchState(0));
+}
+
+setInterval(updateLoop, 1000/24);
